+++ Sources/TextFinder.h
@@ -79,7 +79,7 @@
 
 @interface NSString (NSStringTextFinding)
 
-- (NSRange)findString:(NSString *)string selectedRange:(NSRange)selectedRange options:(unsigned)mask wrap:(BOOL)wrapFlag;
+- (NSRange)findString:(NSString *)string selectedRange:(NSRange)selectedRange options:(NSUInteger)mask wrap:(BOOL)wrapFlag;
 
 @end
